Use onEliminatePlayer to end games
I realized it's still possible to use renaming to cheat a player out of a win. I added onEliminatePlayer so that the player count is checked every time a player is removed (alongside onRemovePlayer), which means the game should end as soon as there is 1 player left; this should never let a game end with no winner.